refactor: 1 (kernel) — paths, marker/version, fs all extracted; cycles eliminated.
